backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
chromadb
sentence-transformers
groq
python-dotenv
